refactor: fix class name and file name as per python styling
